refactor: common 폴더 병합
